✨ Create dependabot.yml
Added a new dependabot.yml file to the repository. This configuration file will enable Dependabot to check for updates of the dependencies in our project on a regular basis. This will help us keep our project up-to-date with the latest versions of libraries, ensuring we benefit from recent bug fixes and improvements.